update tests of create_app; minor change setup.cfg
