+++ setup.cfg
@@ -13,7 +13,6 @@ omit =
     settings_user.py
     settings_development.py
     settings_production.py
-    settings_user.py
 
 [coverage:report]
 ignore_errors = True
